% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_estimator.R
\name{base_estimator}
\alias{base_estimator}
\title{Construct a Base Estimator}
\usage{
base_estimator(fit_fn, predict_fn = stats::predict)
}
\arguments{
\item{fit_fn}{a function taking only two parameters, the predictor variables
\code{x} and a vector of outcomes \code{y}. This function should return an
object which has an implementation of \code{\link[stats]{predict}}.}

\item{predict_fn}{a function taking two parameters, the model \code{object}
and the predictor variables \code{newdata}. \code{stats::predict} is
used by default.}
}
\description{
This is a helper to bundle up the model fitting and prediction functions.
All the models in this package expect fit and predict functions which take
only data as arguments. \code{x} and \code{y} for fitting and only \code{x}
for prediction.
}
\examples{
\dontrun{
  library(randomForest)

  rf_fit <- function(x, y) {
    randomForest(x, y, ntree = 200)
  }

  rf_predict <- function(object, newdata) {
    stats::predict(object, newdata, type = "prob")[, 2]
  }

  rf <- base_estimator(rf_fit, rf_predict)
}

}
