% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv_predict}
\alias{cv_predict}
\alias{cv_predict.default}
\alias{cv_predict.formula}
\title{Out of Sample Predictions Using Cross Valiation}
\usage{
cv_predict(x, model, folds, ...)

\method{cv_predict}{default}(x, y, tmt, model, folds = 3, ...)

\method{cv_predict}{formula}(x, data, model, folds = 3, ...)
}
\arguments{
\item{x}{a \code{data.frame} or matrix with predictor variables measured prior to
treatment or unaffected by treatment. Alternatively, this can be a model
formula: \code{outcome ~ x1 + x2 | treatment}}

\item{model}{a \code{hete_model} to use for making predictions.
are passed to the \code{model} fitting function.}

\item{folds}{the number of folds to use for cros validation.}

\item{\dots}{extra parameters passed to \code{model}.}

\item{y}{a vector of outcomes.}

\item{tmt}{a vector indicating which units received treatment.}

\item{data}{a \code{data.frame} containing the variables specified in
\code{x} when using a formula to specify the model.}
}
\description{
Out of Sample Predictions Using Cross Valiation
}
