% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hete_split.R
\name{hete_split}
\alias{hete_split}
\alias{hete_split.default}
\alias{hete_split.formula}
\title{Split Estimators for Control and Treatment}
\usage{
hete_split(x, ...)

\method{hete_split}{default}(x, y, tmt, est, tmt_est = est, ctl_est = est,
  ...)

\method{hete_split}{formula}(x, data, est, tmt_est = est, ctl_est = est,
  ...)
}
\arguments{
\item{x}{a \code{data.frame} or matrix with predictor variables measured prior to
treatment or unaffected by treatment. Alternatively, this can be a model
formula: \code{outcome ~ x1 + x2 | treatment}}

\item{...}{ignored.}

\item{y}{a vector of outcomes.}

\item{tmt}{a vector indicating which units received treatment.}

\item{est, tmt_est, ctl_est}{an estimator to use for modeling the treatment
effect. This must be a function which takes two arguments, \code{x} and
\code{y} and returns an object which has an implementation of
\code{\link[stats]{predict}}. The \code{predict} implementation must return
a single vector with the estimated outcome or probability of success in the
case of binary outcomes. By default the same estimator, \code{est} is used
for both models in the meta-estimator. You may optionally select a
different estimator for each models. \code{tmt_est} and \code{ctl_est} are
used to estimate the response in the treatment and control groups
respectively.}

\item{data}{a \code{data.frame} containing the variables specified in
\code{x} when using a formula to specify the model.}
}
\value{
A \code{hete_split} object.
}
\description{
Split Estimators for Control and Treatment
}
\examples{
\dontrun{
library(tidyverse)
data(gotv)

df <- gotv \%>\%
  filter(treatment \%in\% c("Control", "Neighbors")) \%>\%
  mutate(treatment = ifelse(treatment == "Control", 0, 1))

m <- hete_split(voted ~ . | treatment, data = df, est = random_forest)
p <- predict(m, df)

uc <- uplift(df$treatment, df$voted, p)
plot(uc)

}
}
