% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hete_x.R
\name{hete_x}
\alias{hete_x}
\alias{hete_x.default}
\alias{hete_x.formula}
\title{X-Learner Meta-Estimator}
\usage{
hete_x(x, ...)

\method{hete_x}{default}(x, y, tmt, est, ...)

\method{hete_x}{formula}(x, data, est, ...)
}
\arguments{
\item{x}{a \code{data.frame} or matrix with predictor variables measured prior to
treatment or unaffected by treatment. Alternatively, this can be a model
formula: \code{outcome ~ x1 + x2 | treatment}}

\item{...}{ignored.}

\item{y}{a vector of outcomes.}

\item{tmt}{a vector indicating which units received treatment.}

\item{est}{an estimator to use for modeling the treatment effect. This must
be a function which takes two arguments, \code{x} and \code{y} and returns
an object which has an implementation of \code{\link[stats]{predict}}. The
\code{predict} implementation must return a single vector with the
estimated outcome or probability of success in the case of binary outcomes.}

\item{data}{a \code{data.frame} containing the variables specified in
\code{x} when using a formula to specify the model.}
}
\value{
A \code{hete_x} object.
}
\description{
X-Learner Meta-Estimator
}
\examples{
\dontrun{
library(tidyverse)
data(gotv)

df <- gotv \%>\%
  filter(treatment \%in\% c("Control", "Neighbors")) \%>\%
  mutate(treatment = ifelse(treatment == "Control", 0, 1))

m <- hete_x(voted ~ . | treatment, data = df, est = random_forest)
p <- predict(m, df)

uc <- uplift(df$treatment, df$voted, p)
plot(uc)

}
}
\seealso{
\url{https://arxiv.org/abs/1706.03461}
}
