% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_estimator.R
\name{random_forest}
\alias{random_forest}
\title{Wrapper for ranger}
\usage{
random_forest(x, y)
}
\arguments{
\item{x}{a \code{data.frame} or matrix with predictor variables.}

\item{y}{a vector of outcomes or labels.}
}
\description{
Wrapper for ranger
}
